Rename interactor to match naming convention. Remove old controller and point everything to the dialog controller.
